refactor: dtos and models
