+++ src/models/chatModel.ts
@@ -1,31 +1,7 @@
-import { default as Chat, default as IChatDocument } from "@src/dtos/chatDto";
-import { Schema, model } from "mongoose";
-import notiSchema from "./notiModel";
+import { model } from "mongoose";
 
-const chatModel = new Schema<IChatDocument>(
-  {
-    chatName: { type: String, trim: true },
-    isGroupChat: { type: Boolean, default: false },
-    users: [{ type: Schema.Types.ObjectId, ref: "User" }],
-    latestMessage: {
-      type: Schema.Types.ObjectId,
-      ref: "Message",
-    },
-    groupAdmin: { type: Schema.Types.ObjectId, ref: "User" },
-    isDeleted: {
-      type: Schema.Types.Boolean,
-      index: true,
-      default: false
-    },
-    noti: [notiSchema],
-    topNoti: { type: Number, default: null },
-    joinDates: {
-      type: Schema.Types.Mixed,
-      default: []
-    }
-  },
-  { timestamps: true }
-);
+import IChatDocument from "@src/models/interfaces/IChat";
+import ChatSchema from "@src/models/schemas/chatSchema";
 
-const Chat = model("Chat", chatModel);
+const Chat = model<IChatDocument>("Chat", ChatSchema);
 export default Chat;
