+++ src/models/interfaces/IUser.ts
