+++ src/models/schemas/pictureModel.ts
@@ -1,0 +1,14 @@
+import IPictureDocument from "@src/models/interfaces/IPicture";
+
+import { Schema } from "mongoose";
+
+const PictureSchema = new Schema<IPictureDocument>(
+    {
+      chat: { type: Schema.Types.ObjectId, ref: "Chat" },
+      uploadedBy: { type: Schema.Types.ObjectId, ref: "User" },
+      url: { type: String, trim: true },
+    },
+    { timestamps: true }
+);
+
+export default PictureSchema;
