+++ src/models/schemas/userSchema.ts
@@ -1,0 +1,27 @@
+import { Schema } from "mongoose";
+
+import IUserDocument from "@src/models/interfaces/IUser";
+import bcrypt from "bcrypt";
+
+const UserSchema = new Schema<IUserDocument>(
+  {
+    nickname: { type: String, required: true },
+    pic: {
+      type: String,
+      default:
+        "https://icon-library.com/images/anonymous-avatar-icon/anonymous-avatar-icon-25.jpg",
+    },
+    isAdmin: {
+      type: Boolean,
+      required: true,
+      default: false,
+    },
+  },
+  { timestamps: true }
+);
+
+UserSchema.methods.matchPassword = async function (enteredPassword: string) {
+  return await bcrypt.compare(enteredPassword, this.password);
+};
+
+export default UserSchema
