+++ src/models/userModel.ts
@@ -1,28 +1,7 @@
-import { Schema, model } from "mongoose";
+import { model } from "mongoose";
 
-import IUserDocument from "@dtos/userDto";
-import bcrypt from "bcrypt";
+import IUserDocument from "@src/models/interfaces/IUser";
+import UserSchema from "./schemas/userSchema";
 
-const userSchema = new Schema<IUserDocument>(
-  {
-    nickname: { type: String, required: true },
-    pic: {
-      type: String,
-      default:
-        "https://icon-library.com/images/anonymous-avatar-icon/anonymous-avatar-icon-25.jpg",
-    },
-    isAdmin: {
-      type: Boolean,
-      required: true,
-      default: false,
-    },
-  },
-  { timestamps: true }
-);
-
-userSchema.methods.matchPassword = async function (enteredPassword: string) {
-  return await bcrypt.compare(enteredPassword, this.password);
-};
-
-const User = model("User", userSchema);
+const User = model<IUserDocument>("User", UserSchema);
 export default User;
